refactor cli and eplus into submodule
