+++ frads/cli/test.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+def test1():
+    print("test")
+    return "test"
