+++ pyproject.toml
@@ -22,8 +22,8 @@ license = {"file"= "license.txt"}
 requires-python = ">= 3.8"
 
 [project.scripts]
-gen = "frads.cli:gen"
-mrad = "frads.cli:mrad"
+gen = "frads.cli.gen:gen"
+mrad = "frads.cli.mrad:mrad"
 
 [tool.setuptools.dynamic]
 version = {attr = "frads.__version__"}
